requirements.txt
ultralytics
streamlit
opencv-python
python-dotenv
mediapipe
groq